clean up files and folder structure
