+++ LBHTenancyAPI/Gateways/Contacts/Dynamics365RestApiContactsGateway.cs
@@ -3,6 +3,8 @@ using System.Text;
 using System.Threading;
 using System.Threading.Tasks;
 using LBH.Data.Domain;
+using LBHTenancyAPI.Infrastructure.Dynamics365.Client.Factory;
+using LBHTenancyAPI.Infrastructure.Dynamics365.Exceptions;
 using LBHTenancyAPI.UseCases.Contacts.Models;
 using Newtonsoft.Json;
