+++ LBHTenancyAPI/Infrastructure/Dynamics365/Client/Factory/Dynamics365ClientFactory.cs
@@ -1,11 +1,8 @@
-using System;
-using System.Net.Http;
-using System.Threading;
 using System.Threading.Tasks;
 using LBHTenancyAPI.Infrastructure.Dynamics365.Authentication;
 using LBHTenancyAPI.Settings.CRM;
 
-namespace LBHTenancyAPI.Gateways.Contacts
+namespace LBHTenancyAPI.Infrastructure.Dynamics365.Client.Factory
 {
     public class Dynamics365ClientFactory : IDynamics365ClientFactory
     {
@@ -31,44 +28,6 @@ namespace LBHTenancyAPI.Gateways.Contacts
             _client.AddDefaultHeader("Authorization", "Bearer " + accessToken);
 
             return _client;
-        }
-
-    }
-
-    public interface IHttpClient
-    {
-        void AddDefaultHeader(string key, string value);
-        void SetBaseUrl(string baseUrl);
-        Task<HttpResponseMessage> GetAsync(string url, CancellationToken cancellationToken);
-    }
-
-    public class RestfulHttpClient : IHttpClient
-    {
-        private readonly HttpClient _client;
-        public RestfulHttpClient()
-        {
-            _client = new HttpClient();
-        }
-
-        public RestfulHttpClient(HttpMessageHandler messageHandler)
-        {
-            _client = new HttpClient(messageHandler);
-        }
-
-        public void AddDefaultHeader(string key, string value)
-        {
-            _client.DefaultRequestHeaders.Add(key, value);
-        }
-
-        public void SetBaseUrl(string baseUrl)
-        {
-            _client.BaseAddress = new Uri(baseUrl);
-        }
-
-        public async Task<HttpResponseMessage> GetAsync(string url, CancellationToken cancellationToken)
-        {
-            var response = await _client.GetAsync(url, cancellationToken).ConfigureAwait(false);
-            return response;
         }
     }
 }
