+++ LBHTenancyAPI/Infrastructure/Dynamics365/Client/Factory/IDynamics365ClientFactory.cs
@@ -1,15 +1,9 @@
-using System.Net.Http;
 using System.Threading.Tasks;
 
-namespace LBHTenancyAPI.Gateways.Contacts
+namespace LBHTenancyAPI.Infrastructure.Dynamics365.Client.Factory
 {
     public interface IDynamics365ClientFactory
     {
         Task<IHttpClient> CreateClientAsync();
     }
-
-
-
-
-
 }
