+++ LBHTenancyAPI/LBHTenancyAPI.csproj
@@ -4,7 +4,7 @@
     <DockerComposeProjectPath>../docker-compose.dcproj</DockerComposeProjectPath>
   </PropertyGroup>
   <PropertyGroup Condition="'$(Configuration)|$(Platform)'=='Debug|AnyCPU'">
-    <DocumentationFile>\TenancyApi\LBHTenancyAPI\LBHTenancyAPI.xml</DocumentationFile>
+    <DocumentationFile>LBHTenancyAPI.xml</DocumentationFile>
       <NoWarn>$(NoWarn);1591</NoWarn>
   </PropertyGroup>
   <ItemGroup>
