+++ LBHTenancyAPI/LBHTenancyAPI.xml
@@ -1,0 +1,62 @@
+<?xml version="1.0"?>
+<doc>
+    <assembly>
+        <name>LBHTenancyAPI</name>
+    </assembly>
+    <members>
+        <member name="T:LBHTenancyAPI.Gateways.Arrears.Impl.ArrearsActionDiaryGateway">
+            <summary>
+            ArrearsActionDiaryGateway marshalls calls to the Database for reads and Web Service for writes
+            </summary>
+        </member>
+        <member name="F:LBHTenancyAPI.Gateways.Arrears.Impl.ArrearsActionDiaryGateway._actionDiaryService">
+            <summary>
+            WCF Service Interface which allows us to create action diary entries
+            </summary>
+        </member>
+        <member name="T:LBHTenancyAPI.Gateways.Arrears.Impl.ArrearsAgreementGateway">
+            <summary>
+            ArrearsActionDiaryGateway marshalls calls to the Database for reads and Web Service for writes
+            </summary>
+        </member>
+        <member name="F:LBHTenancyAPI.Gateways.Arrears.Impl.ArrearsAgreementGateway._actionDiaryService">
+            <summary>
+            WCF Service Interface which allows us to create action diary entries
+            </summary>
+        </member>
+        <member name="T:LBHTenancyAPI.Infrastructure.API.APIResponse`1">
+            <summary>
+            API Response wrapper for all API responses
+            If a request has been successful this will be denoted by the statusCode
+                Then the 'data' property will be populated
+            If a request has not been successful denoted
+                Then the Error property will be populated
+            </summary>
+            <typeparam name="T"></typeparam>
+        </member>
+        <member name="T:LBHTenancyAPI.Infrastructure.API.IRequest">
+            <summary>
+            Request Interface
+            </summary>
+        </member>
+        <member name="M:LBHTenancyAPI.Infrastructure.API.IRequest.Validate``1(``0)">
+            <summary>
+            Encapsulates the validation by making the request responsible for validating itself
+            </summary>
+            <typeparam name="T"></typeparam>
+            <param name="request"></param>
+            <returns></returns>
+        </member>
+        <member name="T:LBHTenancyAPI.Infrastructure.Validation.RequestValidationResponse">
+            <summary>
+            Encapsulates a valid response using some sort of validation extensions
+            We want this so we can swap out validation tools and still keep a standard response
+            </summary>
+        </member>
+        <member name="T:LBHTenancyAPI.UseCases.ArrearsActions.CreateArrearsActionDiaryUseCase">
+            <summary>
+            Use Case for creating Arrears actions diary entry
+            </summary>
+        </member>
+    </members>
+</doc>
