+++ LBHTenancyAPI/UseCases/Contacts/GetContactsForTenancyUseCase.cs
@@ -1,7 +1,6 @@
 using System.Linq;
 using System.Threading;
 using System.Threading.Tasks;
-using LBHTenancyAPI.Gateways.Arrears;
 using LBHTenancyAPI.Gateways.Contacts;
 using LBHTenancyAPI.Infrastructure.Exceptions;
 using LBHTenancyAPI.UseCases.Contacts.Models;
